# storage/api-docs/api-docs.json
{
    "openapi": "3.0.0",
    "info": {
        "title": "L5 OpenApi",
        "description": "L5 Swagger OpenApi description",
        "contact": {
            "email": "darius@matulionis.lt"
        },
        "license": {
            "name": "Apache 2.0",
            "url": "https://www.apache.org/licenses/LICENSE-2.0.html"
        },
        "version": "1.0.0"
    },
    "paths": {
        "/": {
            "get": {
                "description": "Home page",
                "operationId": "58658e34a5c8bd71c4a0afed6de10d20",
                "responses": {
                    "default": {
                        "description": "Welcome page"
                    }
                }
            }
        },
        "/api/lista-pre-cadastro": {
            "get": {
                "tags": [
                    "Pre Cadastros"
                ],
                "summary": "Lista todos os jovens cadastrados",
                "description": "Lista todos os jovens cadastrados",
                "operationId": "eaed7ef09fef0e2bfc94608b11d02cfa",
                "parameters": [
                    {
                        "name": "name",
                        "in": "query",
                        "description": "name",
                        "required": false
                    },
                    {
                        "name": "email",
                        "in": "query",
                        "description": "email",
                        "required": false
                    },
                    {
                        "name": "page",
                        "in": "query",
                        "description": "Page Number",
                        "required": false
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "content": {
                            "application/json": {}
                        }
                    }
                }
            }
        },
        "/api/gravar-pre-cadastro": {
            "post": {
                "tags": [
                    "Pre Cadastros"
                ],
                "summary": "Grava dados do jovem aprendiz",
                "description": "Grava dados do jovem aprendiz",
                "operationId": "c94cd7f56952cdfa9b17532e8b9746a8",
                "parameters": [
                    {
                        "name": "name",
                        "in": "query",
                        "description": "name",
                        "required": false
                    },
                    {
                        "name": "email",
                        "in": "query",
                        "description": "email",
                        "required": false
                    },
                    {
                        "name": "page",
                        "in": "query",
                        "description": "Page Number",
                        "required": false
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "content": {
                            "application/json": {}
                        }
                    }
                }
            }
        }
    },
    "components": {
        "schemas": {
            "PreCadastroController": {}
        }
    },
    "tags": [
        {
            "name": "Pre Cadastros",
            "description": "Pre Cadastros"
        }
    ]
}